requirements.txt
librosa==0.7.0
numpy==1.17.1
opencv-contrib-python>=4.2.0.34
opencv-python==4.1.0.25
torch==1.1.0
torchvision==0.3.0
tqdm==4.45.0
numba==0.48
pyaudio==0.2.11
imageio==2.3.0
imageio-ffmpeg==0.4.3
pandas==0.23.4
PyYAML==5.1
scikit-image==0.14.0
ffmpeg-python==0.2.0
google-cloud-texttospeech==2.2.0
google-cloud-speech==2.0.1
six==1.16.0
requests==2.25.1